Email_VM/requirements.txt
requests>=2.32.4
ibm-watsonx-ai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
ibm-watsonx-orchestrate